requirements.txt
# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
yfinance==0.2.33

# Data Processing
pandas==2.1.3
numpy==1.26.2

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2

# Google AI
google-generativeai==0.3.1

# Database
sqlalchemy==2.0.23

# Utils
python-multipart==0.0.6

# Additional packages
openai==1.12.0
python-telegram-bot==20.8
aiofiles==23.2.1 